Made one small change.
On the main menu, the button that navigates to the DoctorHome activity had the onclick method "doctorHome" in the Java file. But in the XML, the onClick attribute was set as < android:onClick="docHome" >. "docHome" is the name you gave the Intent in the Java page, but not the name of the actual method (the "public void doctorHome" part). The onClick attribute in the XML needs to be set to the method name, not the Intent name.